fix(ci): move the sandbox helper out of the mutation scope
`.hooks/lib/` is a mutation scope, so `repo-root.mjs` living there put its
own literals under Stryker instrumentation. The suites that compute their
repo root through it then resolve that root out of a mutated conditional,
which the guard-pair scanner correctly declines to resolve — the
`repo-root-1` shard aborted on a red dry run for exactly that reason.

The helper has no library consumer: only test suites import it. It belongs
in `test/helpers/`, which no scope mutates.

- move `repo-root.mjs` and its suite to `test/helpers/`
- repoint the six importers
- drop the `repo-root` shard entry

`test/helpers/repo-root.test.mjs` stays claimed by a runner: it has no
dot-segment, so `node --test` default discovery finds it, which
`test/test-runner-partition.test.mjs` asserts.

Verified: the derivation probe still reports
`config/auto-resolve-regen-rules.json`, `.hooks/pre-push` and
`THREAT-MODEL.md` as FOUND; `run-mutation-dry-run.sh` exits 0 over all 39
mutated files; 3556 node tests and 2472 pytest tests pass.
