+++ test/threat-model-layers.test.mjs
@@ -22,7 +22,7 @@ import { describe, it } from "node:test";
 // `git rev-parse --show-toplevel` used to resolve this and appeared to work,
 // because the sandbox sits inside the checkout and git walks up out of it — a
 // silent escape that held only while `tempDirName` stayed at its default.
-import { repoRoot } from "../.hooks/lib/repo-root.mjs";
+import { repoRoot } from "./helpers/repo-root.mjs";
 
 const read = (relative) => readFileSync(join(repoRoot, relative), "utf8");
